refactor thegraph
