+++ .maintain/rust-docs/index.html
@@ -21,7 +21,7 @@
   <li><a href="./assistants/relay-e2e">assistants/relay-e2e</a></li>
   <li><a href="./assistants/relay-s2s">assistants/relay-s2s</a></li>
   <li><a href="./assistants/subquery">assistants/subquery</a></li>
-  <li><a href="./assistants/thegraph-liketh">assistants/thegraph-liketh</a></li>
+  <li><a href="./assistants/thegraph">assistants/thegraph</a></li>
   <li><a href="./traits/bridge-e2e">traits/bridge-e2e</a></li>
   <li><a href="./traits/bridge-s2s">traits/bridge-s2s</a></li>
 </ul>
