+++ bridges/pangoro-goerli/bridge/Cargo.toml
@@ -56,5 +56,5 @@ client-contracts      = { path = "../../../frame/assistants/client-contracts" }
 client-pangoro        = { path = "../../../frame/assistants/client-pangoro", features = [ "bridge-ethv2" ] }
 client-beacon         = { path = "../../../frame/assistants/client-beacon" }
 relay-e2e             = { path = "../../../frame/assistants/relay-e2e" }
-thegraph-liketh       = { path = "../../../frame/assistants/thegraph-liketh" }
+thegraph              = { path = "../../../frame/assistants/thegraph", features = [ "bridge-ethv2" ] }
