--- frame/assistants/thegraph-liketh/src/graphql/generic/transactions_last.query.graphql
@@ -1,15 +1,0 @@
-query LastTransaction {
-  transactionEntities(
-    first: 1
-    orderBy: blockNumber
-    orderDirection: desc
-  ) {
-    id
-    origin
-    blockNumber
-    blockHash
-    txHash
-    txIndex
-    txType
-  }
-}
