--- frame/assistants/thegraph-liketh/src/graphql/generic/transactions_page_gte_from.query.graphql
@@ -1,18 +1,0 @@
-query TransactionPage($from: Int!, $first: Int!) {
-  transactionEntities(
-    where: {
-      blockNumber_gte: $from
-    }
-    first: $first
-    orderBy: blockNumber
-    orderDirection: asc
-  ) {
-    id
-    origin
-    blockNumber
-    blockHash
-    txHash
-    txIndex
-    txType
-  }
-}
