--- frame/assistants/thegraph-liketh/src/lib.rs
@@ -1,5 +1,0 @@
-pub mod component;
-pub mod config;
-pub mod error;
-pub mod graph;
-pub mod types;
