+++ frame/assistants/thegraph/src/graphql/generic/message_accepted_event.query.graphql
