+++ frame/assistants/thegraph/src/lib.rs
@@ -1,0 +1,10 @@
+pub use self::component::*;
+pub use self::config::*;
+pub use self::error::*;
+pub use self::thegraph::*;
+
+mod component;
+mod config;
+mod error;
+mod thegraph;
+pub mod types;
