+++ frame/assistants/thegraph/src/types/mod.rs
@@ -1,13 +1,13 @@
+#[cfg(feature = "bridge-ethv2")]
+pub use self::bridge_ethv2::*;
 #[allow(unused_imports)]
 pub use self::graphql::*;
 pub use self::mark::*;
-pub use self::message_events::*;
 #[allow(unused_imports)]
 pub use self::resp::*;
-pub use self::tx::*;
 
+#[cfg(feature = "bridge-ethv2")]
+mod bridge_ethv2;
 mod graphql;
 mod mark;
-mod message_events;
 mod resp;
-mod tx;
